# openspec/specifications/apis/openapi.yaml
openapi: 3.0.3
info:
  title: AI Tool Usage Tracker API
  version: 1.0.0
  description: |
    REST API for the AI Tool Usage Tracker platform (Phase 1 MVP).

    **Authentication:** JWT Bearer token except `/auth/login`, `/auth/refresh`, and `/health`.

    **Errors:** RFC 7807 Problem Details (`application/problem+json`).

    **Traceability:** FR-PLT-001, ADR-012, NFR-SEC-003/004/006.

    ## RBAC Roles and Access

    | Role | Scope |
    |---|---|
    | `super_admin` | Full access to all endpoints and data across the entire organisation |
    | `team_admin` | Access to members, tools (read), teams (read), thresholds, uploads, collectors (read), and dashboard analytics — all scoped to managed teams only |
    | `finance_viewer` | Access to reports (read) and `GET /dashboard/my-usage` only |
    | `auditor` | Access to audit logs and `GET /dashboard/my-usage` only |
    | `team_member` | Access to `GET /dashboard/my-usage` only |
  contact:
    name: Platform Team
  license:
    name: Proprietary
servers:
  - url: https://api.ai-tool-tracker.example/api/v1
    description: Production
  - url: https://staging-api.ai-tool-tracker.example/api/v1
    description: Staging
  - url: http://localhost:8000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /teams/{teamId}/members:
    get:
      tags: [Teams]
      summary: List team members
      description: '**Roles:** `super_admin`, `team_admin`.'
      operationId: listTeamMembers
      parameters:
        - $ref: './components/parameters.yaml#/TeamIdPath'
        - $ref: './components/parameters.yaml#/Limit'
        - $ref: './components/parameters.yaml#/Cursor'
      responses:
        '200':
          description: Team members
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: './components/schemas.yaml#/TeamMember'
                  meta:
                    $ref: './components/schemas.yaml#/PaginationMeta'
        '401':
          $ref: './components/responses.yaml#/Unauthorized'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
        '404':
          $ref: './components/responses.yaml#/NotFound'
    post:
      tags: [Teams]
      summary: Add team member
      description: '**Roles:** `super_admin`, `team_admin` (administered teams only).'
      operationId: addTeamMember
      parameters:
        - $ref: './components/parameters.yaml#/TeamIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas.yaml#/TeamMemberAddRequest'
      responses:
        '201':
          description: Member added
          content:
            application/json:
              schema:
                $ref: './components/schemas.yaml#/TeamMember'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
        '404':
          $ref: './components/responses.yaml#/NotFound'
        '409':
          $ref: './components/responses.yaml#/Conflict'
  /teams/{teamId}/members/{userId}:
    delete:
      tags: [Teams]
      summary: Remove team member
      description: Historical usage retained. **Roles:** `super_admin`, scoped `team_admin`.
      operationId: removeTeamMember
      parameters:
        - $ref: './components/parameters.yaml#/TeamIdPath'
        - $ref: './components/parameters.yaml#/MemberUserId'
      responses:
        '204':
          description: Member removed
        '403':
          $ref: './components/responses.yaml#/Forbidden'
        '404':
          $ref: './components/responses.yaml#/NotFound'
...
  /dashboard/tokens:
    get:
      tags: [Dashboard]
      summary: Total token usage widget
      description: |
        **Performance target:** p95 <= 3s (NFR-PER-001).

        **Roles:** `super_admin`, `team_admin`. `team_admin` receives data scoped to
        managed teams only. `finance_viewer`, `auditor`, and `team_member` are blocked.
      operationId: getDashboardTokens
      parameters:
        - $ref: './components/parameters.yaml#/DateFrom'
        - $ref: './components/parameters.yaml#/DateTo'
        - $ref: './components/parameters.yaml#/TeamId'
      responses:
        '200':
          description: Token usage aggregate
          content:
            application/json:
              schema:
                $ref: './components/schemas.yaml#/TokenUsageWidget'
              examples:
                default:
                  $ref: './components/examples.yaml#/TokenWidgetExample'
        '400':
          $ref: './components/responses.yaml#/BadRequest'
        '401':
          $ref: './components/responses.yaml#/Unauthorized'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
  /dashboard/cost:
    get:
      tags: [Dashboard]
      summary: Cost overview widget
      description: |
        **Roles:** `super_admin`, `team_admin`. `team_admin` receives data scoped to
        managed teams only. `finance_viewer`, `auditor`, and `team_member` are blocked.
      operationId: getDashboardCost
      parameters:
        - $ref: './components/parameters.yaml#/DateFrom'
        - $ref: './components/parameters.yaml#/DateTo'
        - $ref: './components/parameters.yaml#/TeamId'
      responses:
        '200':
          description: Cost overview
          content:
            application/json:
              schema:
                $ref: './components/schemas.yaml#/CostOverviewWidget'
              examples:
                default:
                  $ref: './components/examples.yaml#/CostWidgetExample'
        '401':
          $ref: './components/responses.yaml#/Unauthorized'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
  /dashboard/usage-by-tool:
    get:
      tags: [Dashboard]
      summary: Usage by tool widget
      description: |
        **Roles:** `super_admin`, `team_admin`. `team_admin` receives data scoped to
        managed teams only. `finance_viewer`, `auditor`, and `team_member` are blocked.
      operationId: getDashboardUsageByTool
      parameters:
        - $ref: './components/parameters.yaml#/DateFrom'
        - $ref: './components/parameters.yaml#/DateTo'
        - $ref: './components/parameters.yaml#/TeamId'
      responses:
        '200':
          description: Per-tool usage breakdown
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: './components/schemas.yaml#/UsageByToolItem'
                  last_updated_at:
                    type: string
                    format: date-time
        '401':
          $ref: './components/responses.yaml#/Unauthorized'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
  /dashboard/usage-by-team:
    get:
      tags: [Dashboard]
      summary: Usage by team widget
      description: |
        **Roles:** `super_admin`, `team_admin`. `team_admin` receives data scoped to
        managed teams only. `finance_viewer`, `auditor`, and `team_member` are blocked.
      operationId: getDashboardUsageByTeam
      parameters:
        - $ref: './components/parameters.yaml#/DateFrom'
        - $ref: './components/parameters.yaml#/DateTo'
      responses:
        '200':
          description: Team comparison data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        team_id:
                          type: string
                          format: uuid
                        team_name:
                          type: string
                        total_tokens:
                          type: integer
                          format: int64
                        estimated_cost:
                          type: number
        '401':
          $ref: './components/responses.yaml#/Unauthorized'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
  /dashboard/top-consumers:
    get:
      tags: [Dashboard]
      summary: Top consumers widget
      description: |
        **Roles:** `super_admin`, `team_admin`. `team_admin` receives data scoped to
        managed teams only. `finance_viewer`, `auditor`, and `team_member` are blocked.
      operationId: getDashboardTopConsumers
      parameters:
        - $ref: './components/parameters.yaml#/DateFrom'
        - $ref: './components/parameters.yaml#/DateTo'
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
        - name: entity
          in: query
          schema:
            type: string
            enum: [teams, users]
      responses:
        '200':
          description: Ranked top consumers
          content:
            application/json:
              schema:
                type: object
                properties:
                  teams:
                    type: array
                    items:
                      $ref: './components/schemas.yaml#/TopConsumerItem'
                  users:
                    type: array
                    items:
                      $ref: './components/schemas.yaml#/TopConsumerItem'
        '401':
          $ref: './components/responses.yaml#/Unauthorized'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
  /dashboard/alerts:
    get:
      tags: [Dashboard]
      summary: Active alerts widget
      description: |
        **Roles:** `super_admin`, `team_admin`. `team_admin` receives data scoped to
        managed teams only. `finance_viewer`, `auditor`, and `team_member` are blocked.
      operationId: getDashboardAlerts
      parameters:
        - $ref: './components/parameters.yaml#/TeamId'
      responses:
        '200':
          description: Active threshold breaches
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: './components/schemas.yaml#/ActiveAlertSummary'
        '401':
          $ref: './components/responses.yaml#/Unauthorized'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
  /dashboard/trends:
    get:
      tags: [Dashboard]
      summary: Trend analysis widget
      description: |
        **Roles:** `super_admin`, `team_admin`. `team_admin` receives data scoped to
        managed teams only. `finance_viewer`, `auditor`, and `team_member` are blocked.
      operationId: getDashboardTrends
      parameters:
        - $ref: './components/parameters.yaml#/DateFrom'
        - $ref: './components/parameters.yaml#/DateTo'
        - name: granularity
          in: query
          required: true
          schema:
            $ref: './components/schemas.yaml#/TrendGranularity'
        - $ref: './components/parameters.yaml#/TeamId'
        - $ref: './components/parameters.yaml#/ToolId'
      responses:
        '200':
          description: Usage trend time series
          content:
            application/json:
              schema:
                type: object
                properties:
                  granularity:
                    $ref: './components/schemas.yaml#/TrendGranularity'
                  data:
                    type: array
                    items:
                      $ref: './components/schemas.yaml#/TrendPoint'
        '401':
          $ref: './components/responses.yaml#/Unauthorized'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
  /dashboard/my-usage:
    get:
      tags: [Dashboard]
      summary: Personal usage widget
      description: |
        Returns the caller's own usage data. All authenticated roles permitted.

        Admins may pass `user_id` to inspect another user's data when authorized.
      operationId: getDashboardMyUsage
      parameters:
        - $ref: './components/parameters.yaml#/DateFrom'
        - $ref: './components/parameters.yaml#/DateTo'
        - $ref: './components/parameters.yaml#/UserId'
      responses:
        '200':
          description: Personal usage summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_tokens:
                    type: integer
                    format: int64
                  estimated_cost:
                    type: number
                  by_tool:
                    type: array
                    items:
                      $ref: './components/schemas.yaml#/UsageByToolItem'
        '401':
          $ref: './components/responses.yaml#/Unauthorized'
  /dashboard/daily-breakdown:
    get:
      tags: [Dashboard]
      summary: Single-day usage breakdown by team and user
      description: |
        **Roles:** `super_admin`, `team_admin`. `team_admin` receives data scoped to
        managed teams only. `finance_viewer`, `auditor`, and `team_member` are blocked.
      operationId: getDashboardDailyBreakdown
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date-time
          description: Calendar day (ISO 8601 UTC)
        - $ref: './components/parameters.yaml#/TeamId'
        - $ref: './components/parameters.yaml#/ToolId'
      responses:
        '200':
          description: Team and user usage for one day
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      required: [team_id, team_name, tokens, cost, users]
                      properties:
                        team_id:
                          type: string
                          format: uuid
                        team_name:
                          type: string
                        tokens:
                          type: integer
                          format: int64
                        cost:
                          type: number
                        users:
                          type: array
                          items:
                            type: object
                            required: [user_id, user_name, tokens, cost]
                            properties:
                              user_id:
                                type: string
                                format: uuid
                              user_name:
                                type: string
                              tokens:
                                type: integer
                                format: int64
                              cost:
                                type: number
        '401':
          $ref: './components/responses.yaml#/Unauthorized'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from POST /auth/login. Required on all endpoints except /auth/login, /auth/refresh, /health.